requirements.txt
fastapi[standard]
numpy
scikit-learn
imbalanced-learn
python-dotenv
